Divide split api.serializers in chamber_of_deputies.serializers and core.serializers
